requirements.txt
pandas
sentence-transformers
scikit-learn
faiss-cpu
nltk
requests
beautifulsoup4
numpy
tqdm